requirements.txt
# Django
Django==5.1.2
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
django-filter==23.5

# Configuration
python-decouple==3.8

# Logging
structlog==24.1.0

# CORS Headers
django-cors-headers==4.3.1

# Static Files
whitenoise==6.6.0

# Extensions
django-extensions==3.2.3

# Development Tools
django-debug-toolbar==4.3.0

# Google Search API
google-api-python-client==2.108.0

# Celery & Redis
celery==5.3.4
redis==5.0.1

# HTTP Requests
requests==2.31.0

# Video Downloading
yt-dlp==2023.11.16

# Production Server
gunicorn==21.2.0

